feat: WIP extracting from epub

+++ extract-from-epub.js
@@ -1,15 +1,16 @@
-import fs from 'fs';
+import fs from "fs";
 import path from "path";
-import unzipper from 'unzipper';
-import {convert} from 'html-to-text';
-import glob from 'fast-glob';
+import unzipper from "unzipper";
+import cheerio from "cheerio";
+import { convert } from "html-to-text";
+import glob from "fast-glob";
 
 async function findHTMLFiles(directory) {
-  const filePatterns = ['**/*.html', '**/*.xhtml'];
+  const filePatterns = ["**/*.html", "**/*.xhtml"];
   const options = {
     cwd: directory,
     absolute: true,
-    onlyFiles: true
+    onlyFiles: true,
   };
 
   const htmlFiles = await glob(filePatterns, options);
@@ -18,19 +19,23 @@ async function findHTMLFiles(directory) {
 
 async function extractTextFromEPUB(epubFilePath) {
   try {
-    const tempDirectory = './epub_temp';
+    const tempDirectory = "./epub_temp";
     if (!fs.existsSync(tempDirectory)) {
       fs.mkdirSync(tempDirectory);
     }
 
-    const epubFilename = path.basename(epubFilePath)
+    const epubFilename = path.basename(epubFilePath);
     const epubTxtDirectory = path.resolve(epubFilename + "_txt");
 
-    fs.rmdirSync(epubTxtDirectory, { recursive: true }); 
+    const epubTxtDirectoryExists = fs.existsSync(epubTxtDirectory);
+    if (epubTxtDirectoryExists) {
+      fs.rmdirSync(epubTxtDirectory, { recursive: true });
+    }
 
     await fs.promises.mkdir(epubTxtDirectory);
 
-    await fs.createReadStream(epubFilePath)
+    await fs
+      .createReadStream(epubFilePath)
       .pipe(unzipper.Extract({ path: tempDirectory }))
       .promise();
 
@@ -38,45 +43,69 @@ async function extractTextFromEPUB(epubFilePath) {
 
     const chapterTexts = await Promise.all(
       htmlFiles.map(async (htmlFile) => {
-        const htmlContent = await fs.promises.readFile(htmlFile, 'utf8');
-        const plainText = convert(htmlContent, {wordwrap: 0, selectors: [
-          {selector: 'a', options: { ignoreHref: true }},
-          {selector: 'h1', options: { uppercase: false, prefix: 'Heading', }},
-          {selector: 'h2', options: { uppercase: false, prefix: 'Heading' }},
-          {selector: 'h3', options: { uppercase: false, prefix: 'Heading' }},
-          {selector: 'h4', options: { uppercase: false, prefix: 'Heading' }},
-          {selector: 'img', format: 'skip' },
-        ]});
+        const htmlContent = await fs.promises.readFile(htmlFile, "utf8");
+        // process the html with cheerio
+        const $ = cheerio.load(htmlContent);
+        const selectorsToRemove = ["sup"];
+
+        selectorsToRemove.forEach((selector) => {
+          $(selector).remove();
+        });
+
+        const processedHtml = $.html();
+
+        const plainText = convert(processedHtml, {
+          wordwrap: 0,
+          selectors: [
+            { selector: "a", options: { ignoreHref: true } },
+            {
+              selector: "h1",
+              options: { uppercase: false, prefix: "Heading" },
+            },
+            {
+              selector: "h2",
+              options: { uppercase: false, prefix: "Heading" },
+            },
+            {
+              selector: "h3",
+              options: { uppercase: false, prefix: "Heading" },
+            },
+            {
+              selector: "h4",
+              options: { uppercase: false, prefix: "Heading" },
+            },
+            { selector: "img", format: "skip" },
+          ],
+        });
         return plainText;
-      })
+      }),
     );
 
     await Promise.all(
       chapterTexts.map(async (chapterText, index) => {
-        const txtFilePath = path.join(epubTxtDirectory, `chapter_${index}.txt`)
+        const txtFilePath = path.join(epubTxtDirectory, `chapter_${index}.txt`);
         await fs.promises.writeFile(txtFilePath, chapterText);
-      })
-    )
+      }),
+    );
 
-    fs.rmdirSync(tempDirectory, { recursive: true });
+    // fs.rmdirSync(tempDirectory, { recursive: true });
 
     return chapterTexts;
   } catch (error) {
-    console.error('Error extracting text from EPUB:', error);
+    console.error("Error extracting text from EPUB:", error);
     return [];
   }
 }
 
 // Example usage:
-const epubFilePath = './test-epub/a_game_of_our_own.epub';
+const epubFilePath = "./test-epub/1835.epub";
 
 async function main() {
-
-  const chapterTexts = await extractTextFromEPUB(epubFilePath)
+  const chapterTexts = await extractTextFromEPUB(epubFilePath);
   chapterTexts.forEach((chapterText, index) => {
     console.log(`Chapter ${index + 1}`);
     console.log(`Chars: ${chapterText.length}`);
-    console.log(chapterText.slice(0,1000));
+    console.log(chapterText.slice(0, 1000));
   });
 }
